Fixed connective carousel transition
When transitioning between slides from the standard image carousel, the modal carousel would not move to the next slide.  This is now fixed. Each time a slide transistions in a particular image carousel the associated modal carousel will shift as well. This allows the user to click on any image displayed and the modal window will pull up that image.